+++ max_div/_cli/__init__.py
@@ -1,5 +1,6 @@
 from ._cli import *
 from ._cmd_benchmark import *
 from ._cmd_benchmark_internal import *
+from ._cmd_benchmark_platform import *
 from ._cmd_benchmark_solver import *
 from ._cmd_numba_status import *